add comments, fixed frct_quit()

+++ src/hooks.c
@@ -6,7 +6,7 @@
 /*   By: jmuhlber <jmuhlber@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/28 19:18:27 by jmuhlber          #+#    #+#             */
-/*   Updated: 2024/04/02 14:15:44 by jmuhlber         ###   ########.fr       */
+/*   Updated: 2024/04/02 16:13:17 by jmuhlber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,12 @@
 
 static void	runfrct(t_fractol *frct);
 
+/**
+ * Handles keyboard input for fractal navigation and scaling.
+ * 
+ * @param keydata Keyboard data containing the pressed key.
+ * @param params void pointer to the fractol struct.
+ */
 void	keyaction(mlx_key_data_t keydata, void *params)
 {
 	t_fractol	*frct;
@@ -37,6 +43,13 @@ void	keyaction(mlx_key_data_t keydata, void *params)
 		return ;
 }
 
+/**
+ * Handles mouse actions for fractal zooming.
+ * 
+ * @param a unused.
+ * @param b Mouse scroll direction (zoom in, zoom out).
+ * @param params void pointer to the fractol struct.
+ */
 void	mouseaction(double a, double b, void *params)
 {
 	t_fractol	*frct;
@@ -60,6 +73,13 @@ void	mouseaction(double a, double b, void *params)
 	runfrct(frct);
 }
 
+/**
+ * Resizes the fractal window and adjusts image dimensions accordingly.
+ * 
+ * @param a New width of the window.
+ * @param b New height of the window.
+ * @param params void pointer to the fractol struct.
+ */
 void	resizeaction(int a, int b, void *params)
 {
 	t_fractol	*frct;
@@ -69,7 +89,7 @@ void	resizeaction(int a, int b, void *params)
 	frct = (t_fractol *)params;
 	frct->width = a;
 	frct->height = b;
-	img_ratio = (double)frct->width / (double)frct->height;
+	img_ratio = frct->width / frct->height;
 	ratio = (double)a / (double)b;
 	if (ratio > img_ratio)
 		frct->width = b * img_ratio;
@@ -79,6 +99,11 @@ void	resizeaction(int a, int b, void *params)
 	runfrct(frct);
 }
 
+/**
+ * Frees and exits the program properly with no errors.
+ * 
+ * @param (void)*params void pointer to the fractol struct.
+ */
 void	quitaction(void *params)
 {
 	t_fractol	*frct;
@@ -87,6 +112,12 @@ void	quitaction(void *params)
 	frct_quit(frct, FQ_OK);
 }
 
+/**
+ * Runs the selected fractal calculation.
+ * 
+ * @param frct The fractal data struct.
+ 
+ */
 static void	runfrct(t_fractol *frct)
 {
 	if (frct->fractal_type == 1)
